pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vigil"
version = "0.2.0"
description = "LLM Production Safety Platform — standalone, self-contained. No external canari-llm, breakpoint-ai, or canari-forensics dependencies."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer",
    "httpx>=0.25",
    "pyahocorasick>=2.0",
    "rich>=13.0",
]
keywords = ["llm", "safety", "canari", "breakpoint", "forensics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
vigil = "vigil.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.3",
]
reports = [
    "fpdf2>=2.7",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"vigil.attacks" = ["*.bp.json"]
"vigil.breakpoint.config" = ["*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
